Change comment. Separate methods
